ci(coverity): fix Coverity scans (#5266)
Starting in version 2024.12, the "cov-build" command is only
supposed to be used for "build capture", while the "filesystem capture"
used for non-compiled languages is only available via Coverity CLI
(the "coverity" command).

The GH Action we use does not yet support that, so introduce a
workaround by tricking the Action to run the correct command via command
injection.

Fixes #5254.

Signed-off-by: Alex T. <frozen.and.blue@gmail.com>